requirements.txt
asgiref
Django
django-extensions
djangorestframework
pillow
sqlparse
tzdata
graphviz